+++ src/main/java/model/Activity.java
@@ -7,12 +7,12 @@ import java.util.Date;
 public class Activity {
 
     String activityName;
-    float hoursPerWeek;
+    double hoursPerWeek;
     LocalDate startDate;
     LocalDate endDate;
     int activityID;
 
-    public Activity(String activityName, float hoursPerWeek, LocalDate startDate, LocalDate endDate, int activityID){
+    public Activity(String activityName, double hoursPerWeek, LocalDate startDate, LocalDate endDate, int activityID){
         this.activityName = activityName;
         this.hoursPerWeek = hoursPerWeek;
         this.startDate = startDate;
